Fix Arrow Error config option type
Arrow Error option was implicitly converted from int to float type
when passed to EntityTippedArrow (IProjectile) shoot method.

See https://maven.thiakil.com/forge-1.12-javadoc/net/minecraft/entity/IProjectile.html#shoot-double-double-double-float-float-

	There is a random variable to the trajectory of an arrow,
	given by this.rand.nextGaussian() * 0.0075 * (double)inaccuracy
	for the x, y and z coordinate.

From https://minecraft.gamepedia.com/Arrow
See also https://github.com/SlimeKnights/TinkersConstruct/issues/1797

+++ src/main/java/funwayguy/epicsiegemod/core/ESM_Settings.java
@@ -37,7 +37,7 @@ public class ESM_Settings
 	
 	//Skeleton
 	public static int SkeletonDistance;
-	public static int SkeletonAccuracy;
+	public static float SkeletonAccuracy;
 	
 	//Zombie
 	public static boolean ZombieInfectious;
